+++ automation/taskcluster/decisionlib.py
@@ -518,6 +518,8 @@ class DockerWorkerTask(Task):
     """
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
+        # We use this specific version because our decision task also runs on this one.
+        # We also use that same version in docker/build.dockerfile
         self.docker_image = "ubuntu:bionic-20180821"
         self.max_run_time_minutes = 30
         self.scripts = []
@@ -525,7 +527,6 @@ class DockerWorkerTask(Task):
         self.caches = {}
         self.features = {}
         self.artifacts = []
-        self.curl_scripts_count = 0
 
     with_docker_image = chaining(setattr, "docker_image")
     with_max_run_time_minutes = chaining(setattr, "max_run_time_minutes")
@@ -574,17 +575,11 @@ class DockerWorkerTask(Task):
         return self
 
     def with_curl_script(self, url, file_path):
-        self.curl_scripts_count += 1
-        n = self.curl_scripts_count
         return self \
-        .with_env(**{
-            "CURL_%s_URL" % n: url,
-            "CURL_%s_PATH" % n: file_path,
-        }) \
         .with_script("""
-            mkdir -p $(dirname "$CURL_{n}_PATH")
-            curl --retry 5 --connect-timeout 10 -Lf "$CURL_{n}_URL" -o "$CURL_{n}_PATH"
-        """.format(n=n))
+            mkdir -p $(dirname {file_path})
+            curl --retry 5 --connect-timeout 10 -Lf {url} -o {file_path}
+        """.format(url=url, file_path=file_path))
 
     def with_curl_artifact_script(self, task_id, artifact_name, out_directory=""):
         return self \
@@ -607,7 +602,6 @@ class DockerWorkerTask(Task):
         return self \
         .with_env(**git_env()) \
         .with_early_script("""
-            git init repo
             cd repo
             git fetch --quiet --tags "$GIT_URL" "$GIT_REF"
             git reset --hard "$GIT_SHA"