+++ components/places/ffi/src/lib.rs
@@ -4,8 +4,7 @@
 
 use ffi_support::{
     define_box_destructor, define_bytebuffer_destructor, define_handle_map_deleter,
-    define_string_destructor, rust_str_from_c, rust_string_from_c, ByteBuffer, ConcurrentHandleMap,
-    ExternError,
+    define_string_destructor, ByteBuffer, ConcurrentHandleMap, ExternError, FfiStr,
 };
 use places::error::*;
 use places::{db::PlacesInterruptHandle, storage, ConnectionType, PlacesApi, PlacesDb};
@@ -42,16 +41,16 @@ lazy_static::lazy_static! {
 /// Instantiate a places API. Returned api must be freed with
 /// `places_api_destroy`. Returns null and logs on errors (for now).
 #[no_mangle]
-pub unsafe extern "C" fn places_api_new(
-    db_path: *const c_char,
-    encryption_key: *const c_char,
+pub extern "C" fn places_api_new(
+    db_path: FfiStr<'_>,
+    encryption_key: FfiStr<'_>,
     error: &mut ExternError,
 ) -> u64 {
     log::debug!("places_api_new");
     APIS.insert_with_result(error, || {
-        let path = ffi_support::rust_string_from_c(db_path);
-        let key = ffi_support::opt_rust_string_from_c(encryption_key);
-        PlacesApi::new(path, key.as_ref().map(|v| v.as_str()))
+        let path = db_path.as_str();
+        let key = encryption_key.as_opt_str();
+        PlacesApi::new(path, key)
     })
 }
 
@@ -60,7 +59,7 @@ pub unsafe extern "C" fn places_api_new(
 /// not destroy, or even close, connections already obtained from this
 /// object. Returns null and logs on errors (for now).
 #[no_mangle]
-pub unsafe extern "C" fn places_connection_new(
+pub extern "C" fn places_connection_new(
     handle: u64,
     conn_type_val: u8,
     error: &mut ExternError,
@@ -118,14 +117,14 @@ pub extern "C" fn places_interrupt(handle: &PlacesInterruptHandle, error: &mut E
 /// Add an observation to the database. The observation is a VisitObservation represented as JSON.
 /// Errors are logged.
 #[no_mangle]
-pub unsafe extern "C" fn places_note_observation(
+pub extern "C" fn places_note_observation(
     handle: u64,
-    json_observation: *const c_char,
+    json_observation: FfiStr<'_>,
     error: &mut ExternError,
 ) {
     log::debug!("places_note_observation");
     CONNECTIONS.call_with_result_mut(error, handle, |conn| {
-        let json = ffi_support::rust_str_from_c(json_observation);
+        let json = json_observation.as_str();
         let visit: places::VisitObservation = serde_json::from_str(&json)?;
         places::api::apply_observation(conn, visit)
     })
@@ -134,36 +133,35 @@ pub unsafe extern "C" fn places_note_observation(
 /// Execute a query, returning a `Vec<SearchResult>` as a JSON string. Returned string must be freed
 /// using `places_destroy_string`. Returns null and logs on errors (for now).
 #[no_mangle]
-pub unsafe extern "C" fn places_query_autocomplete(
+pub extern "C" fn places_query_autocomplete(
     handle: u64,
-    search: *const c_char,
+    search: FfiStr<'_>,
     limit: u32,
     error: &mut ExternError,
 ) -> *mut c_char {
     log::debug!("places_query_autocomplete");
-    CONNECTIONS.call_with_result(error, handle, |conn| {
-        search_frecent(
+    CONNECTIONS.call_with_result(error, handle, |conn| -> places::Result<_> {
+        let res = search_frecent(
             conn,
             SearchParams {
-                search_string: ffi_support::rust_string_from_c(search),
+                search_string: search.into_string(),
                 limit,
             },
-        )
+        )?;
+        Ok(serde_json::to_string(&res)?)
     })
 }
 
 /// Execute a query, returning a URL string or null. Returned string must be freed
 /// using `places_destroy_string`. Returns null if no match is found.
 #[no_mangle]
-pub unsafe extern "C" fn places_match_url(
+pub extern "C" fn places_match_url(
     handle: u64,
-    search: *const c_char,
+    search: FfiStr<'_>,
     error: &mut ExternError,
 ) -> *mut c_char {
     log::debug!("places_match_url");
-    CONNECTIONS.call_with_result(error, handle, |conn| {
-        match_url(conn, ffi_support::rust_string_from_c(search))
-    })
+    CONNECTIONS.call_with_result(error, handle, |conn| match_url(conn, search.as_str()))
 }
 
 #[no_mangle]
@@ -190,7 +188,7 @@ pub unsafe extern "C" fn places_get_visited(
             .iter()
             .enumerate()
             .filter_map(|(idx, &p)| {
-                let s = ffi_support::rust_str_from_c(p);
+                let s = FfiStr::from_raw(p).as_str();
                 url::Url::parse(s).ok().map(|url| (idx, url))
             })
             .collect::<Vec<_>>();
@@ -221,14 +219,10 @@ pub extern "C" fn places_get_visited_urls_in_range(
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn places_delete_place(
-    handle: u64,
-    url: *const c_char,
-    error: &mut ExternError,
-) {
+pub extern "C" fn places_delete_place(handle: u64, url: FfiStr<'_>, error: &mut ExternError) {
     log::debug!("places_delete_place");
     CONNECTIONS.call_with_result(error, handle, |conn| -> places::Result<_> {
-        let url = parse_url(rust_str_from_c(url))?;
+        let url = parse_url(url.as_str())?;
         if let Some(guid) = storage::history::url_to_guid(conn, &url)? {
             storage::history::delete_place_by_guid(conn, &guid)?;
         }
@@ -255,9 +249,9 @@ pub extern "C" fn places_delete_visits_between(
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn places_delete_visit(
+pub extern "C" fn places_delete_visit(
     handle: u64,
-    url: *const c_char,
+    url: FfiStr<'_>,
     timestamp: i64,
     error: &mut ExternError,
 ) {
@@ -265,7 +259,7 @@ pub unsafe extern "C" fn places_delete_visit(
     CONNECTIONS.call_with_result(error, handle, |conn| -> places::Result<_> {
         storage::history::delete_place_visit_at_time(
             conn,
-            &parse_url(rust_str_from_c(url))?,
+            &parse_url(url.as_str())?,
             places::Timestamp(timestamp.max(0) as u64),
         )?;
         Ok(())
@@ -273,19 +267,19 @@ pub unsafe extern "C" fn places_delete_visit(
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn places_wipe_local(handle: u64, error: &mut ExternError) {
+pub extern "C" fn places_wipe_local(handle: u64, error: &mut ExternError) {
     log::debug!("places_wipe_local");
     CONNECTIONS.call_with_result(error, handle, |conn| storage::history::wipe_local(conn))
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn places_run_maintenance(handle: u64, error: &mut ExternError) {
+pub extern "C" fn places_run_maintenance(handle: u64, error: &mut ExternError) {
     log::debug!("places_run_maintenance");
     CONNECTIONS.call_with_result(error, handle, |conn| storage::run_maintenance(conn))
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn places_prune_destructively(handle: u64, error: &mut ExternError) {
+pub extern "C" fn places_prune_destructively(handle: u64, error: &mut ExternError) {
     log::debug!("places_prune_destructively");
     CONNECTIONS.call_with_result(error, handle, |conn| {
         storage::history::prune_destructively(conn)
@@ -293,7 +287,7 @@ pub unsafe extern "C" fn places_prune_destructively(handle: u64, error: &mut Ext
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn places_delete_everything(handle: u64, error: &mut ExternError) {
+pub extern "C" fn places_delete_everything(handle: u64, error: &mut ExternError) {
     log::debug!("places_delete_everything");
     CONNECTIONS.call_with_result(error, handle, |conn| {
         storage::history::delete_everything(conn)
@@ -301,7 +295,7 @@ pub unsafe extern "C" fn places_delete_everything(handle: u64, error: &mut Exter
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn places_get_visit_infos(
+pub extern "C" fn places_get_visit_infos(
     handle: u64,
     start_date: i64,
     end_date: i64,
@@ -318,12 +312,12 @@ pub unsafe extern "C" fn places_get_visit_infos(
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn sync15_history_sync(
+pub extern "C" fn sync15_history_sync(
     handle: u64,
-    key_id: *const c_char,
-    access_token: *const c_char,
-    sync_key: *const c_char,
-    tokenserver_url: *const c_char,
+    key_id: FfiStr<'_>,
+    access_token: FfiStr<'_>,
+    sync_key: FfiStr<'_>,
+    tokenserver_url: FfiStr<'_>,
     error: &mut ExternError,
 ) {
     log::debug!("sync15_history_sync");
@@ -331,11 +325,11 @@ pub unsafe extern "C" fn sync15_history_sync(
         // Note that api.sync returns a SyncPing which we drop on the floor.
         api.sync(
             &sync15::Sync15StorageClientInit {
-                key_id: rust_string_from_c(key_id),
-                access_token: rust_string_from_c(access_token),
-                tokenserver_url: parse_url(rust_str_from_c(tokenserver_url))?,
+                key_id: key_id.into_string(),
+                access_token: access_token.into_string(),
+                tokenserver_url: parse_url(tokenserver_url.as_str())?,
             },
-            &sync15::KeyBundle::from_ksync_base64(rust_str_from_c(sync_key))?,
+            &sync15::KeyBundle::from_ksync_base64(sync_key.as_str())?,
         )?;
         Ok(())
     })