+++ components/support/ffi/src/lib.rs
@@ -60,9 +60,10 @@
 //! Inside the Rust component, you will implement:
 //!
 //! 1. [`IntoFfi`] for all types defined in that crate that you want to return
-//!    over the FFI. For most common cases, the [`implement_into_ffi_by_pointer!`] and
-//!    [`implement_into_ffi_by_json!`] macros will do the job here, however you can see that trait's
-//!    documentation for discussion and examples of implementing it manually.
+//!    over the FFI. For most common cases, the [`implement_into_ffi_by_json!`] and
+//!    [`implement_into_ffi_by_protobuf!`] macros will do the job here, however you
+//!    can see that trait's documentation for discussion and examples of
+//!    implementing it manually.
 //!
 //! 2. Conversion to [`ExternError`] for the error type(s) exposed by that
 //!    rust component, that is, `impl From<MyError> for ExternError`.
@@ -87,20 +88,22 @@
 //!    also use the versions of these functions in the [`abort_on_panic`] module, which
 //!    do as their name suggest.
 //!
-//! Additionally, c strings that are passed in as arguments may be converted to rust strings using
-//! helpers such as [`rust_str_from_c`], [`opt_rust_str_from_c`], [`rust_string_from_c`],
-//! [`opt_rust_string_from_c`], etc.
+//! Additionally, c strings that are passed in as arguments may be represented using [`FfiStr`],
+//! which contains several helpful inherent methods for extracting their data.
 //!
 
 use std::{panic, thread};
 
 mod error;
+mod ffistr;
 pub mod handle_map;
 mod into_ffi;
+#[macro_use]
 mod macros;
 mod string;
 
 pub use crate::error::*;
+pub use crate::ffistr::FfiStr;
 pub use crate::into_ffi::*;
 pub use crate::macros::*;
 pub use crate::string::*;
@@ -125,11 +128,6 @@ pub use crate::handle_map::{ConcurrentHandleMap, Handle, HandleError, HandleMap}
 ///
 /// A few points about the following example:
 ///
-/// - This function *must* be unsafe, as it reads from a raw pointer. If you made it safe, then safe
-///   Rust could cause memory safety violations, which would be very bad! (However, FFI functions
-///   that don't read from raw pointers don't need to be marked `unsafe`! Sadly, most of ours need
-///   to take strings, and so we're out of luck...)
-///
 /// - We need to mark it as `#[no_mangle] pub extern "C"`.
 ///
 /// - We prefix it with a unique name for the library (e.g. `mylib_`). Foreign functions are not
@@ -137,7 +135,7 @@ pub use crate::handle_map::{ConcurrentHandleMap, Handle, HandleError, HandleMap}
 ///   including memory safety issues in some cases.
 ///
 /// ```rust,no_run
-/// # use ffi_support::{ExternError, ErrorCode};
+/// # use ffi_support::{ExternError, ErrorCode, FfiStr};
 /// # use std::os::raw::c_char;
 ///
 /// # #[derive(Debug)]
@@ -149,11 +147,9 @@ pub use crate::handle_map::{ConcurrentHandleMap, Handle, HandleError, HandleMap}
 /// # }
 ///
 /// #[no_mangle]
-/// pub unsafe extern "C" fn mylib_print_string(
-///     // Strings come in as a null terminated C string. This is certainly not ideal but it simplifies
-///     // the "FFI consumer" code, which is trickier code to get right, as it typically has poor
-///     // support for interacting with native libraries.
-///     thing_to_print: *const c_char,
+/// pub extern "C" fn mylib_print_string(
+///     // Strings come in as an `FfiStr`, which is a wrapper around a null terminated C string.
+///     thing_to_print: FfiStr<'_>,
 ///     // Note that taking `&mut T` and `&T` is both allowed and encouraged, so long as `T: Sized`,
 ///     // (e.g. it can't be a trait object, `&[T]`, a `&str`, etc). Also note that `Option<&T>` and
 ///     // `Option<&mut T>` are also allowed, if you expect the caller to sometimes pass in null, but
@@ -163,7 +159,7 @@ pub use crate::handle_map::{ConcurrentHandleMap, Handle, HandleError, HandleMap}
 ///     // You should try to to do as little as possible outside the call_with_result,
 ///     // to avoid a case where a panic occurs.
 ///     ffi_support::call_with_result(error, || {
-///         let s = ffi_support::rust_str_from_c(thing_to_print);
+///         let s = thing_to_print.as_str();
 ///         if s.len() == 0 {
 ///             // This is a silly example!
 ///             return Err(BadEmptyString);
@@ -173,42 +169,9 @@ pub use crate::handle_map::{ConcurrentHandleMap, Handle, HandleError, HandleMap}
 ///     })
 /// }
 /// ```
-///
-/// ## Unwind (panic) Safety
-///
-/// Internally, this function wraps it's argument in a
-/// [`AssertUnwindSafe`](std::panic::AssertUnwindSafe). That means it doesn't attempt to force you
-/// to mark things as [`UnwindSafe`](std::panic::UnwindSafe). Effectively, we're saying that every
-/// caller to this function is automatically panic safe, which is a lie. This is not ideal, but it's
-/// unclear what the right call here would be.
-///
-/// To be clear, making the wrong choice here has no bearing on memory safety, unless there are
-/// exisiting memory safety holes in the code. That means by using `AssertUnwindSafe`, we end up in
-/// a position closer to the position we'd be in if we were working in a language with exceptions,
-/// which typically provides little-to-no assistance in terms of program correctness in the case of
-/// something `throw`ing.
-///
-/// Anyway, if we *were* to require `F: UnwindSafe`, the implementer of the FFI component would need
-/// to use `AssertUnwindSafe` on every FFI binding that wraps a method that needs to call something
-/// on a `&mut T` (note that this is *not* true for `*mut T`, which we want to discourage). The use
-/// of this seems likely to be frequent enough in this FFI that I have an extremely hard time
-/// believing it would be used with consideration, so while the strategy of "assume everything is
-/// panic-safe" is clearly not great, it seems likely to be what happens anyway.
-///
-/// There are, of course, other options:
-///
-/// 1. Abort on panic (e.g. only expose the implementations in `abort_on_panic`), which is bad
-///    for obvious reasons, and seems even worse given our position as libraries.
-/// 2. Poison on panic (as [`std::sync::Mutex`] does, for example). This is a valid option, but
-///    seems wrong for all cases.
-/// 3. Re-initialize on panic (e.g. reopen the DB connection).
-///
-/// 2 and 3 are promising, and allowing users of `ffi-support` to make these choices with a low
-/// amount of boilerplate is something we'd like to investigate in the future, but currently this
-/// is where we've landed.
 pub fn call_with_result<R, E, F>(out_error: &mut ExternError, callback: F) -> R::Value
 where
-    F: FnOnce() -> Result<R, E>,
+    F: panic::UnwindSafe + FnOnce() -> Result<R, E>,
     // It would be nice to only require std::fmt::Debug if the `log_backtraces`
     // feature is on, but there's not really a way to do that in stable rust (at least
     // not in a way that wouldn't add more work for consumers of this lib).
@@ -232,7 +195,7 @@ where
 /// the crate top-level docs for more info.
 pub fn call_with_output<R, F>(out_error: &mut ExternError, callback: F) -> R::Value
 where
-    F: FnOnce() -> R,
+    F: panic::UnwindSafe + FnOnce() -> R,
     R: IntoFfi,
 {
     // We need something that's `Into<ExternError>`, even though we never return it, so just use
@@ -246,23 +209,18 @@ fn call_with_result_impl<R, E, F>(
     abort_on_panic: bool,
 ) -> R::Value
 where
-    F: FnOnce() -> Result<R, E>,
+    F: panic::UnwindSafe + FnOnce() -> Result<R, E>,
     E: Into<ExternError> + std::fmt::Debug,
     R: IntoFfi,
 {
     *out_error = ExternError::success();
-    // It's not ideal to handle unwind safety this way, however I'm not sure we can reasonably
-    // expect the FFI code to think about this in a meaningful way. That said, you cannot cause
-    // memory safety violations by breaking unwind safety (note that this function is not `unsafe`),
-    // short of bugs in unsafe code elsewhere, so this isn't the *worst* thing we could be doing.
-    let res: thread::Result<(ExternError, R::Value)> =
-        panic::catch_unwind(panic::AssertUnwindSafe(|| {
-            init_backtraces_once();
-            match callback() {
-                Ok(v) => (ExternError::default(), v.into_ffi_value()),
-                Err(e) => (e.into(), R::ffi_default()),
-            }
-        }));
+    let res: thread::Result<(ExternError, R::Value)> = panic::catch_unwind(|| {
+        init_backtraces_once();
+        match callback() {
+            Ok(v) => (ExternError::default(), v.into_ffi_value()),
+            Err(e) => (e.into(), R::ffi_default()),
+        }
+    });
     match res {
         Ok((err, o)) => {
             *out_error = err;
@@ -283,6 +241,7 @@ where
 /// that aborts, instead of unwinding, on panic.
 pub mod abort_on_panic {
     use super::*;
+    use std::panic::AssertUnwindSafe;
 
     /// Same as the root `call_with_result`, but aborts on panic instead of unwinding. See the
     /// `call_with_result` documentation for more.
@@ -292,7 +251,7 @@ pub mod abort_on_panic {
         E: Into<ExternError> + std::fmt::Debug,
         R: IntoFfi,
     {
-        super::call_with_result_impl(out_error, callback, true)
+        super::call_with_result_impl(out_error, AssertUnwindSafe(callback), true)
     }
 
     /// Same as the root `call_with_output`, but aborts on panic instead of unwinding. As a result,
@@ -306,7 +265,7 @@ pub mod abort_on_panic {
         let mut dummy = ExternError::success();
         super::call_with_result_impl(
             &mut dummy,
-            || -> Result<_, ExternError> { Ok(callback()) },
+            AssertUnwindSafe(|| -> Result<_, ExternError> { Ok(callback()) }),
             true,
         )
     }